+++ Legacy/classes.php
@@ -85,7 +85,7 @@ if (class_exists('Sys25\RnBase\Typo3Wrapper\Core\CommandLineController')) {
 \class_alias(\Sys25\RnBase\Utility\Misc::class, 'tx_rnbase_util_Misc');
 \class_alias(\Sys25\RnBase\Domain\Collection\BaseCollection::class, 'Tx_Rnbase_Domain_Collection_Base');
 \class_alias(\Sys25\RnBase\Domain\Model\DataInterface::class, 'Tx_Rnbase_Domain_Model_DataInterface');
-\class_alias(\Sys25\RnBase\Domain\Model\DomainInterface::class, 'Tx_Rnbase_Domain_Model_DomainInterface');
+\class_alias(\Sys25\RnBase\Domain\Model\DomainModelInterface::class, 'Tx_Rnbase_Domain_Model_DomainInterface');
 \class_alias(\Sys25\RnBase\Domain\Model\DynamicTableInterface::class, 'Tx_Rnbase_Domain_Model_DynamicTableInterface');
 \class_alias(\Sys25\RnBase\Domain\Model\RecordInterface::class, 'Tx_Rnbase_Domain_Model_RecordInterface');
 \class_alias(\Sys25\RnBase\Domain\Model\RecordInterface::class, 'tx_rnbase_IModel');